requirements.txt
streamlit==0.61.0
numpy==1.18.1
tensorflow-cpu==2.3.0
pillow==6.2.0
